feat: check if user exists, delete it and check if delete successfully

+++ src/main/java/org/example/postusercomment/user/UserService.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -27,7 +28,20 @@ public class UserService {
         userRepository.update(user);
     }
 
-    public void deleteUser(Long id) {
-        userRepository.delete(id);
+    public String deleteUser(Long id) {
+
+        boolean userExists = userRepository.checkIfExistsById(id);
+        if (userExists) {
+            userRepository.delete(id);
+            boolean userDeleted = userRepository.checkIfExistsById(id);
+            if (userDeleted) {
+                return "User with id " + id + " is not deleted.";
+            } else {
+                return "User with id " + id + " successfully deleted";
+            }
+        } else {
+            return "User with id " + id + " does not exist.";
+        }
+
     }
 }
